feat(ui): FeaturePaywall + server-side gating for /planner and /resume (#167)
Page-level gating for the two Pro-only surfaces, with a grandfathered
read-only mode for free users who had data before the gate landed.

New components:

- `components/billing/FeaturePaywall.tsx` — full-page treatment for
  free users who hit a gated surface with no existing data. Distinct
  from the modal `UpgradePromptDialog`: paywalls gate a *feature*, the
  modal gates a *limit*. Copy driven by FEATURE_META so paywall,
  pricing, and upgrade dialog stay in sync.
- `FeatureReadOnlyBanner` — narrower inline variant shown above the
  grandfathered read-only view ("…is now a Pro feature. You can still
  view what you created before, but new changes need a Pro plan.").
- `UpgradeCheckoutButton` — extracted from the monolithic
  UpgradePromptDialog so paywall + banners can embed the same Stripe
  checkout flow without pulling in the dialog chrome.

Page wiring:

- `app/planner/page.tsx` and `app/resume/page.tsx` are now server
  components. They auth, look up the plan, and branch:
  - Pro → render the interactive client as before.
  - Free + has existing rows → render the client in `isReadOnly` mode
    (banner + disabled mutations).
  - Free + no existing rows → render `<FeaturePaywall feature="…" />`.
- Interactive code moved to `PlannerClient.tsx` + `ResumeClient.tsx`
  without behavioural changes for Pro users. Day checkboxes, New
  Plan button, Upload card, and AI question generation are hidden /
  disabled when `isReadOnly`. Delete stays visible — cleanup of
  already-owned data is never a Pro feature.
- `PlanCardMenu` gains an `isReadOnly` prop that hides the Archive
  item but keeps Delete.

Tests: page-level client tests now import `PlannerClient` / `ResumeClient`
directly since the default export is now a server component. Pro-path
behaviour is unchanged; server-level branching is exercised by the API
integration tests plus the `features.test.ts` unit suite.
